Add cpp file for dkstd_curl

+++ dkstd_string.hpp
@@ -69,7 +69,7 @@ inline std::string dkstd::ws2s(std::wstring wstr)
 template<typename ...Args>
 inline std::string dkstd::format_string(const std::string & format, Args ...args)
 {
-    size_t size = std::snprintf(nullptr, 0, format.c_str(), args...) + 1;	// Extra space for '\0'
+    int size = std::snprintf(nullptr, 0, format.c_str(), args...) + 1;	// Extra space for '\0'
     std::unique_ptr<char[]> buf(new char[size]);
     std::snprintf(buf.get(), size, format.c_str(), args...);
     return std::string(buf.get(), buf.get() + size - 1);					// We don't want the '\0' inside
@@ -80,7 +80,7 @@ inline std::string dkstd::format_string(const std::string & format, Args ...args
 template<typename ...Args>
 inline std::wstring dkstd::format_string(const std::wstring & format, Args ...args)
 {
-    size_t size = std::swprintf(nullptr, 0, format.c_str(), args...) + 1;	// Extra space for '\0'
+    int size = std::swprintf(nullptr, 0, format.c_str(), args...) + 1;	// Extra space for '\0'
     std::unique_ptr<wchar_t[]> buf(new wchar_t[size]);
     std::swprintf(buf.get(), size, format.c_str(), args...);
     return std::wstring(buf.get(), buf.get() + size - 1);					// We don't want the '\0' inside
